requirements.txt
Django==5.1.3
django-cors-headers==4.6.0
django-crontab==0.7.1
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf_spectacular==0.28.0